readme: use 'chef gem install' instead of just 'gem install'.
The knife-joyent gem needs to be installed in the chef ruby runtime
environment.

Per the Chef Docs here: https://docs.chef.io/resource_gem_package.html

The chef_gem and gem_package resources are both used to install Ruby
gems. For any machine on which the chef-client is installed, there are
two instances of Ruby. One is the standard, system-wide instance of
Ruby and the other is a dedicated instance that is available only to
the chef-client. Use the chef_gem resource to install gems into the
instance of Ruby that is dedicated to the chef-client. Use the
gem_package resource to install all other gems (i.e. install gems
system-wide).